+++ bbin/pack_lib.sh
@@ -1,19 +1,17 @@
 #!/bin/bash
 
-export CURR_DIR=$(pwd)
-echo $CURR_DIR
-
 export SCRIPT_DIR=$(dirname $(readlink -f $0))
 echo $SCRIPT_DIR
 
-export PROJ_DIR=$SCRIPT_DIR/..
+export PROJ_DIR=$(pwd)
+if [ ! -f "setup.py"]; then
+    echo "Please go to project folder!"
+    exit 1
+fi
 echo $PROJ_DIR
 
 $SCRIPT_DIR/cleanup.sh
 
-cd $PROJ_DIR
 python setup.py bdist_wheel sdist
 
 mv $PROJ_DIR/src/*.egg-info $PROJ_DIR
-
-cd $CURR_DIR
